# packages/BarryKit/Sources/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Barry API",
    "version": "1.0.0",
    "description": "The single supported HTTP contract for Barry applications."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/traits": {
      "get": {
        "operationId": "listTraits",
        "parameters": [
          {
            "name": "identityId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Traits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraitListResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTrait",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTraitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trait",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraitResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "default": "about:blank",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "additionalProperties": false
      },
...
      "TraitListResponse": {
        "type": "object",
        "properties": {
          "traits": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string",
                  "nullable": true
                },
                "access": {
                  "type": "string",
                  "enum": [
                    "read",
                    "readwrite"
                  ]
                },
                "namespaces": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "skills": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "tools": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "instructions": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "bound": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                },
                "boundNames": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "bag": {
                  "type": "string",
                  "nullable": true
                }
              },
              "required": [
                "name",
                "description",
                "access",
                "namespaces"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "traits"
        ],
        "additionalProperties": false
      },
      "TraitResponse": {
        "type": "object",
        "properties": {
          "trait": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "description": {
                "type": "string",
                "nullable": true
              },
              "access": {
                "type": "string",
                "enum": [
                  "read",
                  "readwrite"
                ]
              },
              "namespaces": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "skills": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "tools": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "instructions": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "bound": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {}
              },
              "boundNames": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "bag": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "name",
              "description",
              "access",
              "namespaces"
            ],
            "additionalProperties": false,
            "nullable": true
          }
        },
        "required": [
          "trait"
        ],
        "additionalProperties": false
      },
      "CreateTraitRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "namespaces": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "access": {
            "type": "string",
            "enum": [
              "read",
              "readwrite"
            ]
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "instructions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "boundNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "namespaces",
          "access"
        ],
        "additionalProperties": false
      },
...
    }
  }
}